+++ packages/module-crm/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @brightweblabs/module-crm
 
+## 0.18.3
+
+### Patch Changes
+
+- Updated dependencies [ebdc5fc]
+  - @brightweblabs/core-auth@0.12.1
+  - @brightweblabs/app-shell@0.16.2
+  - @brightweblabs/ui@1.5.5
+  - @brightweblabs/module-orgs@0.7.2
+
+## 0.18.2
+
+### Patch Changes
+
+- Updated dependencies [80915d3]
+  - @brightweblabs/app-shell@0.16.1
+  - @brightweblabs/module-orgs@0.7.1
+
 ## 0.18.1
 
 ### Patch Changes